feat: #1118 add verifyAllWaysTableCellContent command
