+++ mw-webapp/cypress/e2e/allWays.cy.ts
@@ -4,6 +4,7 @@ import {headerSelectors} from "cypress/scopesSelectors/headerSelectors";
 import {navigationMenuSelectors} from "cypress/scopesSelectors/navigationMenuSelectors";
 import {Symbols} from "src/utils/Symbols";
 import {userPersonalSelectors} from "cypress/scopesSelectors/userPersonalDataSelectors";
+import testUserData from "cypress/fixtures/testUserDataFixture.json";
 
 afterEach(() => {
   cy.clearAllStorage();
@@ -32,34 +33,21 @@ describe('NoAuth All Ways scope tests', () => {
 
   it('NoAuth_AllWaysTable_LinkToOwner', () => {
     let actualUserName: string;
-    let targetTableHeaderIndex: number;
-    let linkColumnIndex;
     const targetTableHeader = allWaysPageContent.waysTable.columns.owner.en;
+    const rowIndex = 0;
 
     allWaysSelectors.filterViewBlock.getTableViewButton().click();
 
-
-    allWaysSelectors.allWaysTable.getTableTh().each((th, index) => {
-      const headerText = th.text().trim();
-      if (headerText === targetTableHeader) {
-        targetTableHeaderIndex = index;
-        return false;
-      }
-    });
+    cy.verifyAllWaysTableCellContent(targetTableHeader, allWaysSelectors.allWaysTable.getOwnerLink(), rowIndex);
 
     allWaysSelectors.allWaysTable.getOwnerLink()
-      .first()
+      .eq(rowIndex)
       .then (link => {
         actualUserName = link.text().trim();
         cy.wrap(link).click();
-
-        cy.wrap(link).parents('td').then(td => {
-          linkColumnIndex = td.index();
-          assert.equal(linkColumnIndex, targetTableHeaderIndex);
-        });
       })
       .then(() => {
-        cy.url().should('match', /\/user\/[0-9a-f]{8}-([0-9a-f]{4}-){3}[0-9a-f]{12}$/);
+        cy.url().should('match', new RegExp(testUserData.userUrlPattern));
         userPersonalSelectors.descriptionSection.getName().should('have.text', actualUserName);
       });
   });
